vulkan SSBOs

+++ src/Cinenic.Renderer/OpenGL/GLShaderArrayData.cs
@@ -7,7 +7,7 @@ namespace Cinenic.Renderer.OpenGL {
 	
 	public class GLShaderArrayData<T> : IShaderArrayData<T> {
 		
-		public uint Handle { get; private set; }
+		public ulong Handle { get; private set; }
 		
 		public uint Binding { get; init; }
 
@@ -36,15 +36,15 @@ namespace Cinenic.Renderer.OpenGL {
 				}
 			}
 			
-			_platform.API.NamedBufferData(Handle, Size, dataPtr, VertexBufferObjectUsage.DynamicDraw);
-			_platform.API.BindBufferBase(BufferTargetARB.ShaderStorageBuffer, Binding, Handle);
+			_platform.API.NamedBufferData((uint) Handle, Size, dataPtr, VertexBufferObjectUsage.DynamicDraw);
+			_platform.API.BindBufferBase(BufferTargetARB.ShaderStorageBuffer, Binding, (uint) Handle);
 		}
 
 		public unsafe void Read() {
 			Debug.Assert(Handle != 0);
 
 			void* ptr = null;
-			_platform.API.GetNamedBufferSubData(Handle, 0, Size, ptr);
+			_platform.API.GetNamedBufferSubData((uint) Handle, 0, Size, ptr);
 
 			fixed(void* dataPtr = &_data[0]) {
 				Buffer.MemoryCopy(ptr, dataPtr, Size, Size);
@@ -53,7 +53,7 @@ namespace Cinenic.Renderer.OpenGL {
 
 		public void Dispose() {
 			GC.SuppressFinalize(this);
-			if(Handle != 0) _platform.API.DeleteBuffer(Handle);
+			if(Handle != 0) _platform.API.DeleteBuffer((uint) Handle);
 			Handle = 0;
 		}
 	}
